colados-image-processor/requirements.txt
fastapi==0.118.0
inference_sdk==0.58.1
minio==7.2.18
pika==1.3.2
pydantic==2.11.10
python-dotenv==1.1.1
Requests==2.32.5
SQLAlchemy==2.0.43
uvicorn==0.37.0
